Adds selinux label
For selinux base systems, like fedora, selinux policy does not allow the
container to access the host's binded mounts. That's why we need to
specify the z selinux label in the bind mount options. The same has to
be done in case of the short bind format e.g. -<src>:<dest>:<params>.
check the docker documentation for more information on volumes.