Ensure :latest points to the latest and greatest, not the newest
The old script used to update the 'monetdb/monetdb:latest' tag on
every push.  This means that one day 'monetdb/monetdb:latest' might
point to 'monetdb/monetdb:Sep2022-SP1' and the next day it could
point to 'monetdb/monetdb:Jan2020-SP7'.

Users expect the :latest tag to point to the latest and greatest,
most suitable release, not the most recently released version.

When invoked with tag 'banana', this script pushes
	monetdb/monetdb:banana

When invoked with tag 'Feb2023', this script pushes
	monetdb/monetdb:Feb2023
	monetdb/monetdb:Feb2023-latest
	monetdb/monetdb:latest

When invoked with tag 'Feb2023-SP1', this script pushes
	monetdb/monetdb:Feb2023-SP1
	monetdb/monetdb:Feb2023-latest

(and not monetdb/monetdb:latest)